change duration to be indexed as milliseconds
this will make the SASI on this column contain fewer distinct values, improving write performance.
users won't be discerning within milliseconds about searches they perform in the UI, in regards to the duration field.

duration is always rounded up, so the result list would not skip any results. and duration in the span table would remain accurate to microseconds.

ref: https://github.com/openzipkin/zipkin/pull/1758#discussion_r143885341